requirements.txt
PyYAML==6.0
pandas==1.3.5
pymongo[srv]==4.3.2
matplotlib==3.5.1
scikit-learn==1.0.2
#ipython==7.23.1
nltk==3.7
pretty_confusion_matrix==0.1.1
opencv-python==4.6.0.66
